The *Great Renaming:* gracklepy edition
