+++ doc/source/Contributing.rst
@@ -12,7 +12,7 @@ This section provides a short list of tasks to complete when a new field is intr
    (i) the field is initialized to the appropriate default value.
    (ii) the new field and its value are printed when ``grackle_verbose = 1``
    (iii) the field can be accessed by the functions providing the dynamic access to members of :c:data:`chemistry_data` (see :ref:`dynamic-api`)
-   (iv) the new field is accessible from Pygrackle (because that interface uses the functions providing dynamic access)
+   (iv) the new field is accessible from Gracklepy (because that interface uses the functions providing dynamic access)
 
 2. Update the fortran definition for the ``grackle_chemistry_data`` type (in ``src/clib/grackle_fortran_interface.def``). This type must exactly match the definition of :c:data:`chemistry_data` and is used to integrate Grackle into simulation codes written in Fortran.
