+++ src/python/gracklepy/__config__.py.in
@@ -1,4 +1,4 @@
-# Pygrackle's build system replaces __config__.py.in with __config__.py
+# Gracklepy's build system replaces __config__.py.in with __config__.py
 # -> This practice was inspired by numpy and scipy
 # -> everything other than __version__, is intended for internal purposes
 
@@ -20,7 +20,7 @@ def _report_error(variable_name, nominal_value, allowed_vals=None):
     # we use unicode escape sequence, \u0040, that python automatically
     # converts to the "at sign" to avoid unintentional substitutions
     raise RuntimeError(
-        "There appears to have been an issue when pygrackle's build system "
+        "There appears to have been an issue when gracklepy's build system "
         f"generated `{__file__}` from the `{__file__}.in` template file. In "
         f"particular, \u0040{variable_name}\u0040 was replaced with an "
         f"invalid value of '{nominal_value}'.{allowed_value_descr}"
@@ -34,7 +34,7 @@ if (len(__version__) == 0) or __version__[0] == "@":
 
 
 def _is_editable_installation():
-    """Returns whether pygrackle was built as an editable install"""
+    """Returns whether gracklepy was built as an editable install"""
     state = r"@SKBUILD_STATE@"
     # allowed values are given by
     # https://scikit-build-core.readthedocs.io/en/latest/cmakelists.html#accessing-information
@@ -50,6 +50,6 @@ def _is_editable_installation():
 # specifies how we consumed the grackle library
 _GrackleBuild = Enum("_GrackleBuild", "ExternalClassic ExternalCMake Embedded")
 try:
-    _grackle_build = _GrackleBuild[r"@_PYGRACKLE_CONSUME_MODE@"]
+    _grackle_build = _GrackleBuild[r"@_GRACKLEPY_CONSUME_MODE@"]
 except KeyError as err:
     _report_error("SKBUILD_STATE", err.args[0], list(_GrackleBuild.__members__))
