+++ src/python/gracklepy/grackle_wrapper.pyx
@@ -12,7 +12,7 @@
 ########################################################################
 
 import copy
-from pygrackle.utilities.physical_constants import \
+from gracklepy.utilities.physical_constants import \
     boltzmann_constant_cgs, \
     mass_hydrogen_cgs
