+++ src/python/gracklepy/utilities/data_path.py
@@ -13,15 +13,15 @@
 
 import os
 
-from pygrackle.__config__ import _is_editable_installation
-from pygrackle.utilities.misc import dirname
+from gracklepy.__config__ import _is_editable_installation
+from gracklepy.utilities.misc import dirname
 
 if _is_editable_installation():
-    # Note, this only works with an editable install of pygrackle.
+    # Note, this only works with an editable install of gracklepy.
     _install_dir = dirname(os.path.abspath(__file__), level=5)
     grackle_data_dir = os.path.join(_install_dir, "input")
 else:
     raise RuntimeError(
-        "in non-editable pygrackle installations, like this one, "
+        "in non-editable gracklepy installations, like this one, "
         f"grackle_data_dir cannot be imported from {__file__}."
     )
