+++ src/python/setup.cfg
@@ -1,4 +1,4 @@
 [flake8]
 max-line-length = 999
-exclude = pygrackle/__init__.py,pygrackle/utilities/testing.py
+exclude = gracklepy/__init__.py,gracklepy/utilities/testing.py
 ignore = E111,E121,E122,E123,E124,E125,E127,E129,E131,E201,E202,E211,E221,E222,E227,E228,E241,E301,E203,E225,E226,E231,E251,E261,E262,E265,E266,E302,E303,E402,E502,E701,E731,W292,W293,W391,W503,W504
