+++ src/python/tests/test_chemistry.py
@@ -14,12 +14,12 @@
 import numpy as np
 import os
 
-from pygrackle import \
+from gracklepy import \
     chemistry_data, \
     setup_fluid_container, \
     set_cosmology_units
 
-from pygrackle.utilities.testing import \
+from gracklepy.utilities.testing import \
     random_logscale, \
     assert_rel_equal, \
     assert_array_less
