+++ src/python/tests/test_initialisation.py
@@ -11,10 +11,10 @@ import h5py
 import os
 
 #Chemistry_data struct from grackle
-from pygrackle import chemistry_data
+from gracklepy import chemistry_data
 #Necessary constants from grackle
-from pygrackle.utilities.physical_constants import mass_hydrogen_cgs
-from pygrackle.utilities.testing import assert_allclose
+from gracklepy.utilities.physical_constants import mass_hydrogen_cgs
+from gracklepy.utilities.testing import assert_allclose
 
 #* Function which returns chemistry_data instance with default initialisation settings.
 def get_defChem():
