+++ src/python/tests/test_models.py
@@ -5,9 +5,9 @@ import yt
 
 from numpy.testing import assert_allclose
 
-from pygrackle.__config__ import _is_editable_installation
-from pygrackle.utilities.model_tests import model_sets
-from pygrackle.utilities.testing import run_command
+from gracklepy.__config__ import _is_editable_installation
+from gracklepy.utilities.model_tests import model_sets
+from gracklepy.utilities.testing import run_command
 
 from testing_common import grackle_python_dir
