+++ src/python/tests/testing_common.py
@@ -14,7 +14,7 @@
 
 import os
 
-from pygrackle.utilities.misc import dirname
+from gracklepy.utilities.misc import dirname
 
 grackle_install_dir = dirname(os.path.abspath(__file__), level=4)
 grackle_data_dir = os.path.join(grackle_install_dir, "input")
